Added notifications if servers are joined/left

+++ update.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
 
-rm ./programs/ecco.png
+if pgrep -x "./programs/ecco.png" > /dev/null
+then
+    rm ./programs/ecco.png
+fi
 git pull
